Move code completion methods to new file
